// apps/game-client/src/lib/api/generated/main/model/search-timers-npc-response-dto-output.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * Lootlog API
 * The Lootlog API documentation
 * OpenAPI spec version: 1.0
 */
import type { SearchTimersNpcResponseDtoOutputType } from './search-timers-npc-response-dto-output-type';

export interface SearchTimersNpcResponseDtoOutput {
  npcId: number;
  timerKey: string;
  name: string;
  lvl: number;
  type: SearchTimersNpcResponseDtoOutputType;
  prof: string;
  location: string;
  wt: string | number;
  icon: string;
  /** @nullable */
  latestRespBaseSeconds: number | null;
  /** @nullable */
  latestRespawnRandomness: number | null;
}